% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acc}
\alias{acc}
\title{evaluation metric for binary and multiclass classification}
\usage{
acc(y_true, preds)
}
\description{
evaluation metric for binary and multiclass classification
}
\keyword{internal}

