% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_folds}
\alias{class_folds}
\title{stratified folds (in classification)                      [ detailed information about class_folds in the FeatureSelectionR package ]}
\usage{
class_folds(folds, RESP, shuffle = FALSE)
}
\description{
this function creates stratified folds in binary and multiclass classification
}
\keyword{internal}
