% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EXCEPTIONS_preds}
\alias{EXCEPTIONS_preds}
\title{EXCEPTIONS in predictions [ REGRESSION and CLASSIFICATION ]}
\usage{
EXCEPTIONS_preds(FIT, DATA, regression)
}
\description{
EXCEPTIONS in predictions [ REGRESSION and CLASSIFICATION ]
}
\keyword{internal}
