% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_weka_rerun}
\alias{func_weka_rerun}
\title{secondary function for weka-algos (if re-run = TRUE)}
\usage{
func_weka_rerun(bst_lst)
}
\description{
secondary function for weka-algos (if re-run = TRUE)
}
\keyword{internal}
