% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{function_grid}
\alias{function_grid}
\title{grid function  [ this function takes a grid of parameters, IF the grid includes a list and this list includes vectors then it takes a sample of 1 else if it includes a single value then it returns this value ]}
\usage{
function_grid(grid_params = NULL)
}
\description{
grid function  [ this function takes a grid of parameters, IF the grid includes a list and this list includes vectors then it takes a sample of 1 else if it includes a single value then it returns this value ]
}
\keyword{internal}

