% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length_grid}
\alias{length_grid}
\title{function to get the max. length of the parameter-grid}
\usage{
length_grid(grid_params = NULL)
}
\description{
function to get the max. length of the parameter-grid
}
\keyword{internal}
