% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mse}
\alias{mse}
\title{evaluation metric for regression}
\usage{
mse(y_true, y_pred)
}
\arguments{
\item{y_true}{a numeric vector specifying the response variable}

\item{y_pred}{a numeric vector specifying the predictions}
}
\description{
evaluation metric for regression
}
