% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optimal_grid}
\alias{optimal_grid}
\title{function that will be used if re-run = TRUE}
\usage{
optimal_grid(grid_params = NULL, iter = NULL)
}
\description{
function that will be used if re-run = TRUE
}
\keyword{internal}

