% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repeated_resampling}
\alias{repeated_resampling}
\title{Repeated resampling methods}
\usage{
repeated_resampling(y, method, REPEATS = 1, sample_rate = NULL,
  FOLDS = NULL, seed = 1)
}
\arguments{
\item{y}{the response variable (continuous or factor)}

\item{method}{one of 'bootstrap', 'train_test_split', 'cross_validation'}

\item{REPEATS}{the number of times that the method should be repeated}

\item{sample_rate}{train-data-sample-rate for 'bootstrap' and 'train_test_split'}

\item{FOLDS}{the number of folds for the 'cross_validation' method}

\item{seed}{an integer specifying the RNG}
}
\value{
a list of sublist(s)
}
\description{
Repeated resampling methods
}
\details{
This function is meant for splitting the data using three resampling methods, with the option of multiple repeats.
}
\examples{

# data(Boston, package = 'MASS')
# y = Boston$medv
# res = repeated_resampling(y, 'bootstrap', REPEATS = 2, sample_rate = 0.75, FOLDS = NULL)

}
