% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resampling_methods}
\alias{resampling_methods}
\title{Resampling methods used in the RandomSearch file}
\usage{
resampling_methods(y, method, iter = 1, REPEATS = 1, sample_rate = NULL,
  FOLDS = NULL)
}
\description{
Resampling methods used in the RandomSearch file
}
\keyword{internal}

