% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{second_preds}
\alias{second_preds}
\title{secondary function used in performance measures   [ order of calculation is : first 'y_true' then 'preds' ]}
\usage{
second_preds(sublist, eval_metric)
}
\description{
secondary function used in performance measures   [ order of calculation is : first 'y_true' then 'preds' ]
}
\keyword{internal}
