% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_models.R
\name{subset_mods}
\alias{subset_mods}
\title{subset_mods}
\usage{
subset_mods(perf_meas_OBJ, bst_mods = 5, train_params = FALSE)
}
\arguments{
\item{perf_meas_OBJ}{a resulted object of the performance_measures function}

\item{bst_mods}{the number of models to select (rows of each algorithm)}

\item{train_params}{if the subset should be based on train or test output}
}
\value{
a list of lists
}
\description{
subset the parameter output of a performance_measures object
}
\details{
This function takes the result from the performance_measures function, the number of best performing grid-models and a boolean specifying if the train or test predictions should be taken into account
when subseting the data.frames and it returns a list with the optimal parameters for each algorithm.
}
\examples{

# bst_m = subset_mods(perf_meas_OBJ = perf, bst_mods = 5, train_params = FALSE)
# bst_m
}
