% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.gbm}
\alias{predict.gbm}
\title{work around for bug in gbm 2.1.1 with minor modifications -- PREDICTION-PROBABILITIES when predict(fit, data, type = 'response')}
\usage{
\method{predict}{gbm}(object, newdata, n.trees, type = "link", single.tree = FALSE, ...)
}
\description{
work around for bug in gbm 2.1.1 with minor modifications -- PREDICTION-PROBABILITIES when predict(fit, data, type = 'response')
}
\references{
http://www.samuelbosch.com/2015/09/workaround-ntrees-is-missing-in-r.html
}
\keyword{internal}
