% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_gbm_workaround_probs}
\alias{predict_gbm_workaround_probs}
\title{secondary function [ see previous one ]}
\usage{
predict_gbm_workaround_probs(object, newdata, n.trees, type = "link",
  single.tree = FALSE, ...)
}
\description{
secondary function [ see previous one ]
}
\keyword{internal}

