% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resampling_methods}
\alias{resampling_methods}
\title{Resampling methods}
\usage{
resampling_methods(y, method, iter = 1, REPEATS = 1, sample_rate = NULL,
  FOLDS = NULL)
}
\arguments{
\item{y}{the response variable (continuous or factor)}

\item{method}{one of 'bootstrap', 'train_test_split', 'cross_validation'}

\item{iter}{defaults to 1 and can vary when the resampling_methods function is used internally (as is the case in the random_search_resample function)}

\item{REPEATS}{the number of times that the method should be repeated}

\item{sample_rate}{train-data-sample-rate for 'bootstrap' and 'train_test_split'}

\item{FOLDS}{the number of folds for the 'cross_validation' method}
}
\description{
Three different methods : 'bootstrap', 'train_test_split' and 'cross_validation'
}
\details{
This function is meant for splitting the data using three resampling methods, with the option of multiple repeats.
}
\examples{

# data(Boston, package = 'MASS')
# y = Boston$medv
# res = resampling_methods(y, 'bootstrap', 3, REPEATS = 2, sample_rate = 0.75, FOLDS = NULL)

}

